Move filterTransactions file to a better location
